added tournament controller
