+++ .gitignore
@@ -7,7 +7,7 @@
 /server/config/secrets/prod/prod.decrypt.private.php
 /server/public/bundles/
 /server/var/
-/vendor/
+/server/vendor/
 ###< symfony/framework-bundle ###
 
 ###> phpunit/phpunit ###
